chore: make it so vite doesn't need to import from @ourworldindata/* when running
