+++ vite.config-site.mts
@@ -1,4 +1,4 @@
-import { ViteEntryPoint } from "./site/viteUtils.tsx"
+import { ViteEntryPoint } from "./site/viteConstants.ts"
 import { defineViteConfigForEntrypoint } from "./vite.config-common.mts"
 
 export default defineViteConfigForEntrypoint(ViteEntryPoint.Site)
